debug

+++ scripts/CLUE_GLUE_SuperGLUE_benchmark/run_tasks.sh
@@ -1,29 +1,38 @@
 #!/usr/bin/env bash
 
-wget https://atp-modelzoo-sh.oss-cn-shanghai.aliyuncs.com/tutorial/clue_glue_superglue_benchmark/clue_datasets.tgz
-wget https://atp-modelzoo-sh.oss-cn-shanghai.aliyuncs.com/tutorial/clue_glue_superglue_benchmark/glue_datasets.tgz
-wget https://atp-modelzoo-sh.oss-cn-shanghai.aliyuncs.com/tutorial/clue_glue_superglue_benchmark/superglue_datasets.tgz
-tar -zxf clue_datasets.tgz
-tar -zxf glue_datasets.tgz
-tar -zxf superglue_datasets.tgz
-mkdir datasets
-mv clue_datasets/* datasets
-mv glue_datasets/* datasets
-mv superglue_datasets/* datasets
-rm -rf *_datasets
-rm *.tgz
-
+if [ ! -d ./datasets ];then
+  wget https://atp-modelzoo-sh.oss-cn-shanghai.aliyuncs.com/tutorial/clue_glue_superglue_benchmark/clue_datasets.tgz
+  tar -zxf clue_datasets.tgz
+  mkdir datasets
+  mv clue_datasets/* datasets
+  rm -rf *_datasets
+  rm *.tgz
+fi
 export CUDA_VISIBLE_DEVICES="0"
 #CLUE---> AFQMC, CMNLI, CSL, IFLYTEK, TNEWS
-#GLUE---> CoLA, MRPC, QQP, RTE, SST-2
-#SuperGLUE---> BoolQ, CB, COPA, WiC, WSC
-task_name=CoLA
+task_name=WSC
 
-python main_finetune.py --workerGPU=1 \
+python3 main_finetune.py --workerGPU=1 \
+  --mode="train_and_evaluate_on_the_fly" \
   --task_name=${task_name}  \
   --train_input_fp=datasets/${task_name}/train.csv  \
   --eval_input_fp=datasets/${task_name}/dev.csv  \
-  --pretrain_model_name_or_path=google-bert-base-en  \
+  --pretrain_model_name_or_path=google-bert-base-zh  \
+  --train_batch_size=16  \
+  --num_epochs=2.5  \
+  --model_dir=${task_name}_model_dir  \
+  --learning_rate=1e-5  \
+
+
+predict_checkpoint_path=${task_name}_model_dir/model.ckpt-77
+python3 main_finetune.py --workerGPU=1 \
+  --mode="predict_on_the_fly" \
+  --predict_checkpoint_path=$predict_checkpoint_path \
+  --task_name=${task_name}  \
+  --train_input_fp=datasets/${task_name}/train.csv  \
+  --eval_input_fp=datasets/${task_name}/dev.csv  \
+  --predict_input_fp=datasets/${task_name}/dev.csv  \
+  --pretrain_model_name_or_path=google-bert-base-zh  \
   --train_batch_size=16  \
   --num_epochs=2.5  \
   --model_dir=${task_name}_model_dir  \
